+++ examples/example-3/bow_tie_array_driver.m
@@ -21,6 +21,7 @@ Const = sunem_initialise('bow_tie_array',false);
 Const.runMoMsolver       = true;
 Const.runCBFMsolver      = true;
 Const.runJacobisolver    = true;
+Const.runIFBMoMsolver    = true;
 
 % --------------------------------------------------------------------------------------------------
 % Define input files for extracting FEKO data
@@ -41,11 +42,22 @@ Const.SUNEMjackstrfilename     = 'jack_bow_tie_array.str';
 % --------------------------------------------------------------------------------------------------
 % Define additional program flow constants
 % --------------------------------------------------------------------------------------------------
-% TO-DO: Setup some documentation for this
+% TO-DO: Setup some documentation for this - also assign default values if they are not
+% defined explicitely here.
 Const.no_mutual_coupling_array = false; % Deactivate coupling between domains.
 Const.calcSecMBFs = true;      % For MBF based solvers
 Const.useMBFreduction = true;  % SVD applied after the MBFs are generated to retain an orthonormal set
 Const.MBFthreshold = 1000;     % Threshold used for the SVD reduction of the MBFs
+Const.IFBalg = 7;             % Jacobi iterations (7). Adaptive MBF (14).
+Const.IFB_iterations = 4;      % Number of Jacobi iterations. (TO-DO: Ellaborate special meaning, e.g. -1)
+                               % which then looks at Const.IFB_convergence_threshold_percentage;
+Const.IFB_convergence_threshold_percentage = 1E-3;                                
+Const.IFB_CBFs = -1;           % TO-DO: Recheck this - essentially for the Adaptive MBF the number of MBFs
+                               % to use during each iteration
+Const.IFB_debug = 1;
+Const.cache_Z0_V0 = false;     % Precompute the Z0 and V0 terms
+Const.use_DGFM_start = false;  % Use the DGFM to calculate the initial (0th) solution
+
 
 % --------------------------------------------------------------------------------------------------
 % Read the MoM matrix equation from the file